+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## UNRELEASED
 
+- Add batch upserts & batch recompute for collection & on auto-embedding [#253](https://github.com/hypermodeAI/runtime/pull/253)
 - Add auto-embedding for collection based on text checkpoint [#250](https://github.com/hypermodeAI/runtime/pull/250)
 - Remove extraneous types in graphql schemas [#251] (https://github.com/hypermodeAI/runtime/pull/251)
 - Allow arrays as inputs to host functions [#2521] (https://github.com/hypermodeAI/runtime/pull/252)